refactor: header 리팩토링
