--- src/components/atoms/Header/HeaderLogo.tsx
@@ -1,5 +1,0 @@
-import Logo from "@/assets/main_logo.svg";
-
-export const HeaderLogo = () => {
-  return <img className="w-48 h-48" src={Logo} />;
-};
